actualizacion desarrollo LDR

+++ Prueba_LDR/PWM.h
@@ -6,12 +6,17 @@
 	#include "string.h"
 	#include "stm32f4xx_hal.h"
 	#include "cmsis_os2.h"                          // CMSIS RTOS header file
-	#include "rl_net.h"			// Este deja usar el bool
+	#include <stdbool.h>
+//	#include "rl_net.h"			// Este deja usar el bool
 
 	#define PERIODO 8000-1
 	#define PRESCALER 42-1
 	
+
+	
 	void Init_PWM_Pin(void);
-	void Config_PWM_Pulse(uint8_t pulse, bool PWM_Habilitado);
+	void Config_PWM_Pulse(uint32_t pulse, bool PWM_Habilitado);
 	int Init_ThPWM (void);
+	
+	
 #endif
